Move existing benchmarks to new style.
