+++ benchmark/CMakeLists.txt
@@ -11,24 +11,22 @@ function(dissolve_add_benchmark)
   if(NOT DEFINED DISSOLVE_BENCHMARK_SRC)
     message(FATAL_ERROR "Benchmark registered from directory '${CMAKE_CURRENT_LIST_DIR}' did not set a SRC file.")
   endif(NOT DEFINED DISSOLVE_BENCHMARK_SRC)
-  # set(DISSOLVE_BENCHMARK_USE_TEST_DIRECTORY "${PROJECT_SOURCE_DIR}/benchmarks/data")
 
   # Strip path and extension from source file to get benchmark name
   get_filename_component(benchmark_file ${DISSOLVE_BENCHMARK_SRC} NAME_WE)
   get_filename_component(benchmark_file_dir "${CMAKE_CURRENT_LIST_DIR}" NAME)
-  set(BENCHMARK_NAME "${benchmark_file_dir}_${benchmark_file}")
+  set(BENCHMARK_NAME "benchmark_${benchmark_file_dir}_${benchmark_file}")
   message(STATUS "... Benchmark '${BENCHMARK_NAME}' from ${DISSOLVE_BENCHMARK_SRC} in directory '${CMAKE_CURRENT_LIST_DIR}'")
   message(STATUS "    ... working directory = ${DISSOLVE_BENCHMARK_USE_TEST_DIRECTORY}")
 
   # Register executable target
   add_executable(${BENCHMARK_NAME} ${DISSOLVE_BENCHMARK_SRC})
   target_compile_options(${BENCHMARK_NAME} PRIVATE -O3)
-  target_include_directories(${BENCHMARK_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/src ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR})
 
   # Configure target
   target_include_directories(
-    ${BENCHMARK_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/src ${PROJECT_BINARY_DIR}/src ${PROJECT_SOURCE_DIR} ${CONAN_INCLUDE_DIRS_GTEST}
-                              ${CONAN_INCLUDE_DIRS_PUGIXML}
+    ${BENCHMARK_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/src ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR}
+                              ${PROJECT_SOURCE_DIR}/benchmark ${PROJECT_BINARY_DIR}/benchmark ${CONAN_INCLUDE_DIRS_PUGIXML}
   )
   target_link_libraries(
     ${BENCHMARK_NAME}
@@ -37,3 +35,11 @@ function(dissolve_add_benchmark)
   )
 
 endfunction()
+
+set(DISSOLVE_BENCHMARK_INPUT_FILE_PATH \"${PROJECT_SOURCE_DIR}/benchmark/common/\")
+configure_file(benchmarkConfig.h.in benchmarkConfig.h @ONLY)
+
+# Add benchmark subdirectories
+add_subdirectory(classes)
+add_subdirectory(math)
+add_subdirectory(modules)
