+++ benchmark/benchmarkConfig.h.in
@@ -3,4 +3,4 @@
 */
 #pragma once
 
-inline const char *benchmark_path = @BENCHMARK_PATH@;
+inline const char *benchmarkInputFilePath = @DISSOLVE_BENCHMARK_INPUT_FILE_PATH@;
