--- benchmark/benchmarks.cpp
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-// Copyright (c) 2024 Team Dissolve and contributors
-
-#include "benchmark/benchmark.h"
-
-BENCHMARK_MAIN();
