+++ benchmark/classes/CMakeLists.txt
@@ -1,0 +1,2 @@
+dissolve_add_benchmark(SRC box.cpp)
+dissolve_add_benchmark(SRC region.cpp)
