+++ benchmark/math/CMakeLists.txt
@@ -1,0 +1,4 @@
+dissolve_add_benchmark(SRC array2d.cpp)
+dissolve_add_benchmark(SRC array3d.cpp)
+dissolve_add_benchmark(SRC fourier.cpp)
+dissolve_add_benchmark(SRC interpolator.cpp)
