+++ benchmark/modules/CMakeLists.txt
@@ -1,0 +1,4 @@
+dissolve_add_benchmark(SRC energy.cpp)
+dissolve_add_benchmark(SRC forces.cpp)
+dissolve_add_benchmark(SRC gr.cpp)
+dissolve_add_benchmark(SRC siteRDF.cpp)
